+++ CHANGELOG.md
@@ -10,6 +10,7 @@ this project uses date-based 'snapshot' version identifiers.
 ### Added
 - Force non-Windows diff program to consider all files to be text files
 - Extend PDF-based tests to `dvips`
+- Normalize a new pattern of PDF stream
 
 ### Fixed
 - Restore epoch settings for `dvitopdf()`